backend/pyproject.toml
[project]
name = "control-craft-backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "ibm-watsonx-ai>=1.3.42",
    "ibm-watsonx-orchestrate>=2.3.0",
    "ibmcloudant>=0.11.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
